requirements.txt
python-telegram-bot==21.6
yt-dlp==2025.7.7
aiohttp==3.9.5
